mlb_hr_predictor_pro_ultra/requirements.txt
pandas
numpy
scikit-learn==1.3.0
xgboost==1.7.6
streamlit
python-dotenv
pybaseball
mlbstatsapi
joblib
